refactor: organize codebase in mappings, domain and util folders
