+++ .gitignore
@@ -3,3 +3,4 @@ build
 SECRET.md
 generated
 env
+.DS_Store
