+++ src/util/pool.ts
@@ -1,0 +1,17 @@
+import { BigInt, EthereumBlock, log } from "@graphprotocol/graph-ts";
+import { PoolMeta, poolMetaByIdentifier } from "../poolMetas";
+
+// poolId or address
+export function poolFromIdentifier(id: string): PoolMeta {
+  if (!poolMetaByIdentifier.has(id)) {
+    log.critical("poolMeta not found for identifier {}", [id]);
+  }
+  let poolMeta = poolMetaByIdentifier.get(id);
+  return poolMeta;
+}
+
+export function seniorToJuniorRatio(seniorRatio: BigInt): BigInt {
+  return BigInt.fromI32(10)
+    .pow(27)
+    .minus(seniorRatio);
+}
